midway through state rework
