+++ src/Views/GameBoard/GameBoard.jsx
@@ -1,99 +1,74 @@
 import { useEffect, useState } from "react";
-import { Cloud, Clouds, OrbitControls, Sky, SpotLight, Stars } from "@react-three/drei";
+import { Cloud, Clouds, OrbitControls, Sky } from "@react-three/drei";
 import { Canvas } from "@react-three/fiber";
 import { Physics, RigidBody } from "@react-three/rapier";
 import { Perf } from "r3f-perf";
 import { useGameEngine } from "../../Context/useGameEngine.jsx";
 // Components
 import { UI } from "../../components/Ui/UI.jsx";
+import { GameBoardCells } from "../../components/GameBoardCells/GameBoardCells.jsx";
+import { CitizenRed } from '../../assets/citizens/CitizenRed.jsx';
 // Functions
-import { createGameBoard, tileJump } from "../../../utilities.jsx";
-// 3D components
-import TileD from "../../assets/tiles/tileD.jsx";
-// styling
-import styles from "./GameBoard.module.css";
-//test
-import { tileData } from "./testboarddata.js";
+// import { createGameBoard } from "../../../utilities.jsx";
 import randomTileGenerator from "./utils/randomTileGenerator.js";
 import { checkTilePlacement } from "./verifyFunctions.js";
-import { CitizenRed } from '../../assets/citizens/CitizenRed.jsx';
+// styling
+import styles from "./GameBoard.module.css";
 
 
 const GameBoard = () => {
-  // STATES //
   // TILE
   const tileScale = [0.92, 0.92, 0.92];
   const tileSize = 2;
+  // STATES //
   // CAMERA & ENVIRONMENT
   const [enableRotate, setEnableRotate] = useState(true);
   const [sunPosition, setSunPosition ] = useState([50, 100, 150])
   //NEW TILE
-  // const [newTileArray, setNewTileArray] = useState([]);//all tiles
   const [newTile, setNewTile] = useState();// the new tile mesh thing
-  const [newTileData, setNewTileData] = useState();//the tile object 
+  const [newTileData, setNewTileData] = useState();//the new tile object 
   const [newTileType, setNewTileType ] = useState() // string of tile type
   const [newTilePosition, setNewTilePosition] = useState([12,4,0]);//updates the postion
-  const [newTile2DPosition, setNewTile2DPosition] = useState([]);//updates the position
-  const [releaseTile, setReleaseTile] = useState(false);//mkaes it so you cant click after confirm
-  const [tileRotation, setTileRotation] = useState(0);
-  const [ renderTileArr, setRenderTileArr ] = useState([]);
+  const [newTile2DPosition, setNewTile2DPosition] = useState([]);//updates the 2D tile position
+  const [releaseTile, setReleaseTile] = useState(false);//makes it so you cant click after confirm
+  const [tileRotation, setTileRotation] = useState(0); // sets tile rotation
+  const [ renderTileArr, setRenderTileArr ] = useState([]); // renders 3D models to canvas
   // Citizen
   const [citizenPosition, setCitizenPosition ] = useState([])
   const [ isCitizenPhase, setIsCitizenPhase ] = useState(false)
-  const [ isTilePhase, setIsTilePhase ] = useState(false)
-  // Board
-  // const [boardGameMatrix, setBoardGameMatrix] = useState([
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [tileData], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  //   [[], [], [], [], [], [], [], [], [], [], []],
-  // ]);
-
-
-  // CAMERA
-
-
   const [replaceTile,setReplaceTile]=useState(true)
 
-  
+
 
   const drawEventHandler = async (tileType) => {
     const TileComponent = await import(`../../assets/tiles/tile${tileType}.jsx`);
     const renderNewTile = (
       <RigidBody
-        canSleep={false}
+        canSleep={true}
         position={newTilePosition}
         rotation={[0, tileRotation, 0]}
+        restitution={0}
+        enabledTranslations={[ false, true, false]}
+        enabledRotations={[ false, false, false ]}
       >
         <TileComponent.default scale={tileScale} />
       </RigidBody>
     );
-    console.log(renderNewTile,'render new tile');
     setNewTile(renderNewTile);
   };
   
   const renderTile = async (tileType, position, rotation) => {
-    console.log(position, "FUNC POS");
     if(tileType !== undefined) {
       const TileComponent = await import(`../../assets/tiles/tile${tileType}.jsx`);
       const renderNewTile = (
         <RigidBody
-          canSleep={false}
+          canSleep={true}
           position={position}
           rotation={[0, rotation, 0]}
         >
           <TileComponent.default scale={tileScale} />
         </RigidBody>
       );
-      console.log(renderNewTile,'render new tile');
       return renderNewTile
     }
   };
@@ -114,17 +89,10 @@ const GameBoard = () => {
 
   useEffect(() => {
     console.log(renderTileArr, "RENDER TILE ARR");
-    // setBoardGameMatrix((currBoard) => {
-    //   const newboard = JSON.parse(JSON.stringify(currBoard));
-    //   tileData.grid_id={row:5,column:5}
-    //   newboard[5][5] = [tileData];
-    //   return newboard;
-    // });
+    console.log(boardGameMatrix, "MATRIX");
     boardGameMatrix.forEach((row) =>{
       row.forEach((col) => {
         if(col.length > 0) {
-          // console.log(col[0].grid_id.row, "col[0].grid_id.row");
-          // console.log(col[0].orientation * Math.PI/180, "col[0].orientation");
           const position = [(col[0].grid_id.row -5 ) * 2, 0 , (col[0].grid_id.column - 5) * 2]
           renderTile(col[0].tile_type, position, col[0].orientation * Math.PI/180)
           .then((renderedTile) => {
@@ -137,37 +105,8 @@ const GameBoard = () => {
           })
       }})
     })
-    // const currTileArray = [...newTileArray, newTile]
-    // setNewTileArray(currTileArray)
-    // console.log(boardGameMatrix, "boardGameMatrix useEffect");
-    if(turnPhase === 'Place Tile') {
-      setIsCitizenPhase(false)
-      setIsTilePhase(true)
-    }
-    if(turnPhase === 'Place Citizen') {
-      setIsTilePhase(false)
-      setIsCitizenPhase(true)
-    }
+    console.log(turnPhase);
   }, [turnPhase]);
-
-  // console.log(isCitizenPhase, "isCitizenPhase");
-  // console.log(isTilePhase, "isTilePhase");
-
-  
-  const grid = createGameBoard(
-    boardGameMatrix,
-    tileSize,
-    tileScale,
-    setReleaseTile,
-    setNewTilePosition,
-    setNewTile2DPosition,
-    setNewTile,
-    setNewTileData,
-    turnPhase,
-    isCitizenPhase
-  );
-
-  
 
 
   // RENDERING STARTS HERE //
@@ -176,13 +115,12 @@ const GameBoard = () => {
       <UI
         newTileType={newTileType}
         newTile={newTile}
-        setBoardGameMatrix={setBoardGameMatrix}
+        setNewTile={setNewTile}
         tileRotation={tileRotation}
         setTileRotation={setTileRotation}
         boardGameMatrix={boardGameMatrix}
         checkTilePlacement={checkTilePlacement}
         setReleaseTile={setReleaseTile}
-        setNewTile={setNewTile}
         randomTileGenerator={randomTileGenerator}
         setNewTileData={setNewTileData}
         drawEventHandler={drawEventHandler}
@@ -244,6 +182,19 @@ const GameBoard = () => {
              */}
 
 
+            <GameBoardCells 
+              boardGameMatrix={boardGameMatrix}
+              tileSize={tileSize}
+              tileScale={tileScale}
+              setReleaseTile={setReleaseTile}
+              setNewTilePosition={setNewTilePosition}
+              setNewTile2DPosition={setNewTile2DPosition}
+              setNewTile={setNewTile}
+              setNewTileData={setNewTileData}
+              turnPhase={turnPhase}
+              isCitizenPhase={isCitizenPhase}
+            />
+
             {releaseTile && replaceTile ? newTile : null}
 
             {turnPhase === 'Place Citizen' && citizenPosition.length > 0 ? 
@@ -269,7 +220,7 @@ const GameBoard = () => {
                 <boxGeometry args={[25, 0.5, 25]} />
                 <meshStandardMaterial color="#8f4111" />
               </mesh>
-              {grid}
+              {/* {boardGrid} */}
             </RigidBody>
           </Physics>
           {/* HELPERS */}
